+++ server/src/test/java/org/opensearch/action/ingest/SimulatePipelineRequestParsingTests.java
@@ -85,7 +85,7 @@ public class SimulatePipelineRequestParsingTests extends OpenSearchTestCase {
         when(ingestService.getProcessorFactories()).thenReturn(registry);
     }
 
-    public void testParseUsingPipelineStore(boolean useExplicitType) throws Exception {
+    public void testParseUsingPipelineStore() throws Exception {
         int numDocs = randomIntBetween(1, 10);
 
         Map<String, Object> requestContent = new HashMap<>();
@@ -131,7 +131,7 @@ public class SimulatePipelineRequestParsingTests extends OpenSearchTestCase {
         assertThat(actualRequest.getPipeline().getProcessors().size(), equalTo(1));
     }
 
-    public void innerTestParseWithProvidedPipeline() throws Exception {
+    public void testParseWithProvidedPipeline() throws Exception {
         int numDocs = randomIntBetween(1, 10);
 
         Map<String, Object> requestContent = new HashMap<>();
@@ -149,9 +149,9 @@ public class SimulatePipelineRequestParsingTests extends OpenSearchTestCase {
                         doc.put(field.getFieldName(), value);
                         expectedDoc.put(field.getFieldName(), value);
                     } else {
-                        Integer value = randomIntBetween(1, 1000000);
+                        int value = randomIntBetween(1, 1000000);
                         doc.put(field.getFieldName(), value);
-                        expectedDoc.put(field.getFieldName(), value);
+                        expectedDoc.put(field.getFieldName(), (long) value);
                     }
                 } else if (field == VERSION_TYPE) {
                     String value = VersionType.toString(randomFrom(VersionType.INTERNAL, VersionType.EXTERNAL, VersionType.EXTERNAL_GTE));
@@ -163,9 +163,9 @@ public class SimulatePipelineRequestParsingTests extends OpenSearchTestCase {
                         doc.put(field.getFieldName(), value);
                         expectedDoc.put(field.getFieldName(), value);
                     } else {
-                        Integer value = randomIntBetween(1, 1000000);
+                        int value = randomIntBetween(1, 1000000);
                         doc.put(field.getFieldName(), value);
-                        expectedDoc.put(field.getFieldName(), value);
+                        expectedDoc.put(field.getFieldName(), (long) value);
                     }
                 } else {
                     if (randomBoolean()) {