+++ server/src/test/java/org/opensearch/search/aggregations/pipeline/AvgBucketAggregatorTests.java
@@ -80,15 +80,15 @@ public class AvgBucketAggregatorTests extends AggregatorTestCase {
 
     /**
      * Test for issue #30608.  Under the following circumstances:
-     *
+     * <p>
      * A. Multi-bucket agg in the first entry of our internal list
      * B. Regular agg as the immediate child of the multi-bucket in A
      * C. Regular agg with the same name as B at the top level, listed as the second entry in our internal list
      * D. Finally, a pipeline agg with the path down to B
-     *
+     * <p>
      * BucketMetrics reduction would throw a class cast exception due to bad subpathing.  This test ensures
      * it is fixed.
-     *
+     * <p>
      * Note: we have this test inside of the `avg_bucket` package so that we can get access to the package-private
      * `reduce()` needed for testing this
      */